backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
python-multipart
passlib==1.7.4
bcrypt==3.2.2
python-dotenv
pydantic[email]
pyJWT
openai
pydub
aiofiles
websockets
requests
gtts
cloudinary
pdfminer.six
python-docx